# backend/openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: GnnDEX API
  version: 1.17.0
  description: GnnDEX trading exchange backend API v1 draft.
servers:
  - url: http://localhost:4000/v1
    description: local
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Registered
  /auth/login:
    post:
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                twoFactorCode:
                  type: string
                userAgent:
                  type: string
                ipAddress:
                  type: string
      responses:
        "201":
          description: Auth tokens issued
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                userAgent:
                  type: string
                ipAddress:
                  type: string
      responses:
        "201":
          description: New token pair
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "201":
          description: Revocation result
  /auth/2fa/setup:
    post:
      tags: [auth]
      summary: Setup 2FA TOTP secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "201":
          description: Secret and otpauth URL
  /auth/2fa/enable:
    post:
      tags: [auth]
      summary: Enable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, code]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                code:
                  type: string
      responses:
        "201":
          description: 2FA enabled
  /market/tickers:
    get:
      tags: [market]
      summary: Get ticker snapshots
      parameters:
        - in: query
          name: symbols
          schema:
            type: string
          description: Comma-separated symbols (for example BTC-USDT,ETH-USDT)
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Ticker snapshots
  /market/stream/tickers:
    get:
      tags: [market]
      summary: Stream ticker snapshots (SSE)
      parameters:
        - in: query
          name: symbols
          schema:
            type: string
          description: Comma-separated symbols (for example BTC-USDT,ETH-USDT)
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: intervalMs
          schema:
            type: integer
            minimum: 1000
            maximum: 60000
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
  /market/orderbook/{symbol}:
    get:
      tags: [market]
      summary: Get aggregated orderbook
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Orderbook levels
  /market/candles:
    get:
      tags: [market]
      summary: Get OHLCV candles
      parameters:
        - in: query
          name: symbol
          required: true
          schema:
            type: string
        - in: query
          name: interval
          schema:
            type: string
            enum: ["1m", "5m", "15m", "1h", "4h", "1d"]
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Candle list
...
  /wallet/stream/balances:
    get:
      tags: [wallet]
      summary: Stream wallet balances (SSE)
      parameters:
        - in: query
          name: intervalMs
          schema:
            type: integer
            minimum: 1000
            maximum: 60000
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
  /wallet/ledger:
    get:
      tags: [wallet]
      summary: Get wallet ledger
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entries
  /wallet/withdrawals:
    get:
      tags: [wallet]
      summary: List my withdrawals
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [REQUESTED, REVIEW_PENDING, APPROVED, REJECTED, BROADCASTED, CONFIRMED, FAILED]
      responses:
        "200":
          description: Withdrawal list
    post:
      tags: [wallet]
      summary: Request withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, network, address, amount]
              properties:
                asset:
                  type: string
                network:
                  type: string
                address:
                  type: string
                memo:
                  type: string
                amount:
                  type: string
                fee:
                  type: string
      responses:
        "201":
          description: Withdrawal requested
  /wallet/admin-adjust:
    post:
      tags: [wallet]
      summary: Adjust available balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, asset, amount]
              properties:
                email:
                  type: string
                  format: email
                asset:
                  type: string
                amount:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Updated balance
...
  /orders:
    get:
      tags: [orders]
      summary: List orders with filters and pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: symbol
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [NEW, PARTIALLY_FILLED, FILLED, CANCELED, REJECTED]
        - in: query
          name: statuses
          description: Multi status filter (repeatable or comma-separated)
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum: [NEW, PARTIALLY_FILLED, FILLED, CANCELED, REJECTED]
        - in: query
          name: side
          schema:
            type: string
            enum: [BUY, SELL]
        - in: query
          name: type
          schema:
            type: string
            enum: [MARKET, LIMIT, STOP_LIMIT]
        - in: query
          name: fromCreatedAt
          schema:
            type: string
            format: date-time
          description: createdAt lower bound (UTC ISO-8601)
        - in: query
          name: toCreatedAt
          schema:
            type: string
            format: date-time
          description: createdAt upper bound (UTC ISO-8601)
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [CREATED_AT, PRICE, QUANTITY]
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        "200":
          description: Order list with pagination
    post:
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, side, type, quantity]
              properties:
                symbol:
                  type: string
                  example: BTC-USDT
                side:
                  type: string
                  enum: [BUY, SELL]
                type:
                  type: string
                  enum: [MARKET, LIMIT, STOP_LIMIT]
                price:
                  type: string
                quantity:
                  type: string
      responses:
        "201":
          description: Created order
...
  /orders/{orderId}:
    delete:
      tags: [orders]
      summary: Cancel order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Canceled order
...